feat: added a stub component
